# Use key services and data keys where possible for faster EC operations.