--- a/src/libraries/Native/Unix/System.Security.Cryptography.Native.Apple/CMakeLists.txt
+++ b/src/libraries/Native/Unix/System.Security.Cryptography.Native.Apple/CMakeLists.txt
@@ -9,6 +9,7 @@ set(NATIVECRYPTO_SOURCES
     pal_hmac.c
     pal_keyagree.c
     pal_keychain.c
+    pal_keyservices.c
     pal_random.c
     pal_rsa.c
     pal_sec.c