… src/libraries/Native/Unix/System.Security.Cryptography.Native.Apple/pal_keyservices.c
… src/libraries/Native/Unix/System.Security.Cryptography.Native.Apple/pal_keyservices.h
… src/libraries/System.Security.Cryptography.X509Certificates/src/System.Security.Cryptography.X509Certificates.csproj