@vcsjones 2020-06-18
I guess what I am looking to get input on:

1. Are the performance gains worth the introduced complexity of essentially supporting two different sign / verify paths on macOS?
2. Are the performance gains relevant to where data keys exist? e.g. generated and imported? Or do we expect most keys to come from PFX / keychains?
    1. This may be possible for PFX and in some cases keychain keys. The limiting factor is that the private key needs to exportable. If we can't do that, we can't key it in to key services.
3. Are the performance losses on generate and import acceptable?