@vcsjones 2020-07-10
Going to close this for now, will make a new PR when it makes sense.